+++ CHANGELOG.md
@@ -8,6 +8,8 @@
  - BREAKING: renamed animated-bind component to animated-value. "bind" made sense in the context of early Ember, it's not really a thing people say anymore.
  - BREAKING: we only provide insertedSprites at initial render if you set initialInsertion=true. Received sprites are always still provided, because that's what they're for.
  - BREAKING: moved around many internal modules to make it clear what things are publicy importable. 
+ - BREAKING: rename the default cosine-based easing function from `inAndOut` to `easeInAndOut` for consistency with `easeIn` and `easeOut`.
+
 
 
 # v0.0.1-alpha.0